feat(skeleton): add skeleton component review
