+++ packages/admin-ui/src/index.ts
@@ -34,7 +34,8 @@ export * from './visually-hidden'
 export * from './toast'
 export * from './bulk-actions'
 export * from './filters'
-export * from './components/Skeleton'
+export * from './skeleton'
+
 export * from './components/Label'
 export * from './components/Collapsible'
 export * from './components/CollapsibleGroup'
