+++ packages/admin-ui/src/skeleton/index.tsx
@@ -1,0 +1,1 @@
+export * from './skeleton'
