+++ packages/admin-ui/src/skeleton/skeleton.tsx
@@ -1,9 +1,9 @@
 import type { ComponentPropsWithRef } from 'react'
 import { createComponent, useElement } from '@vtex/admin-ui-react'
 import type { VariantProps } from '@vtex/admin-ui-core'
-import { keyframes } from '@vtex/admin-ui-core'
 
 import * as style from './skeleton.style'
+
 /**
  * Represents a UI that doesn’t contain actual content; instead, it shows the loading elements of a page in a shape similar to actual content.
  * It show users that content is loading, offering a vague preview of how content will look once it fully loads.
@@ -15,23 +15,18 @@ import * as style from './skeleton.style'
  *
  * function Component() {
  *  const { loading, data } = useFetch()
- *  return loading ? <Skeleton csx={{ ... }} /> : <h1>{data}</h1>
+ *  return loading ? <Skeleton /> : <h1>{data}</h1>
  * }
  *
  */
 export const Skeleton = createComponent<'div', SkeletonOptions>((props) => {
   const { shape = 'rect', ...restProps } = props
-  const load = keyframes({
-    '0%': { backgroundPosition: '-200px 0' },
-    '100%': { backgroundPosition: 'calc(200px + 100%) 0' },
-  })
 
   return useElement('div', {
     ...restProps,
     baseStyle: {
       ...style.baseline,
       ...style.variants({ shape }),
-      animation: `${load} 1.2s ease-in-out infinite`,
     },
   })
 })
